Achieving a more fluid feel to the responsiveness.
Ive commented out the widths applied to the .box class as this was making it jump from set width to set width depending on your viewport size. I set a max-width to your container to force it stop scaling at 960px.

I also added a new class to your thumbnail so I could set a max-width on that to stop it forcing the intro module down on to a new line.